renamed ui package to activities
