+++ app/src/main/java/com/example/hobbytracker/activities/Shop.java
@@ -1,4 +1,4 @@
-package com.example.hobbytracker.ui;
+package com.example.hobbytracker.activities;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
