+++ app/src/main/res/layout/activity_achievements.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#EEE9E2"
-    tools:context=".ui.AchievementsActivity">
+    tools:context=".activities.AchievementsActivity">
 
     <FrameLayout
         android:id="@+id/title"
