+++ app/src/main/res/layout/activity_add_hobby.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="#EEE9E2"
     android:orientation="vertical"
-    tools:context=".ui.AddHobbyActivity">
+    tools:context=".activities.AddHobbyActivity">
 
 
     <LinearLayout
