+++ app/src/main/res/layout/activity_shop.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:context=".ui.Shop"
+    tools:context=".activities.Shop"
     android:background="#F0EBE4">
 
     <LinearLayout
